Move PlayerGameLog to separate folder
